refactored and added defaul replicate
